Bug 583822  - CC: ACL issues from CA interface CC doc review
git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1173 c9f7a03b-bd48-0410-a16d-cbbf54688b0b

+++ pki/base/common/src/com/netscape/cms/servlet/admin/RegistryAdminServlet.java
@@ -162,8 +162,12 @@ public class RegistryAdminServlet extends AdminServlet {
                 return;
             getProfileImplConfig(req, resp);
         } else if (op.equals(OpDef.OP_DELETE)) {
+            if (!modifyAuthorize(req, resp))
+                return;
             deleteImpl(req, resp);
         } else if (op.equals(OpDef.OP_ADD)) {
+            if (!modifyAuthorize(req, resp))
+                return;
             addImpl(req, resp);
         } else
             sendResponse(ERROR, INVALID_POLICY_IMPL_OP,
